requirements.txt
requests>=2.27.1
pandas>=1.5.0
mysql-connector-python>=8.0.33